+++ src/renderer/src/store/slices/settingsSlice.ts
@@ -16,6 +16,7 @@ export const defaultGeneralSettings: GeneralSettings = {
   followRedirects: true,
   scrollbarAutoHide: false,
   wrapTabs: true,
+  closeToTray: false,
   spellCheckEnabled: true,
   warnWhenSwitchingThemes: true,
   warnWhenExitingWithUnsavedChanges: true,